refactor(main): changed headless pkg to arguments pkg
