+++ cmd/app/main.go
@@ -1,8 +1,8 @@
 package main
 
 import (
+	"job-visualizer/pkg/arguments"
 	"job-visualizer/pkg/gui"
-	"job-visualizer/pkg/headless"
 	"job-visualizer/pkg/shared"
 	"os"
 )
@@ -11,6 +11,6 @@ func main() {
 	resourceDirectory, err := os.Getwd()
 	shared.CheckError(err)
 	shared.Program.ResourcesDirectory = resourceDirectory
-	isHeadless := headless.CheckCLIArguments()
+	isHeadless := arguments.CheckCLIArguments()
 	gui.RunGUIorHeadless(isHeadless)
 }
